feat(core): XPath criteria, 2xx default success, and retry semantics upgrade
P0-5: xpath criteria evaluate through the existing XpathEvaluator stack
(context expression or response body; non-XML or eval errors fail the
criterion deterministically; pinned versions still validated).

P1-2: operation steps without explicit success criteria now pass on any
2xx response (spec-tooling convention); steps without a recorded status
code keep the previous pass-through.

P1-1: retry semantics upgraded -
- retryAfter accepts decimal seconds (parser optionalNumber)
- HTTP Retry-After header (seconds or RFC7231 date) overrules the
  declared delay when present
- configurable retryBackoffMultiplier scales the delay per upcoming
- attempt number (1.0 = fixed delay, spec default)
- response headers are now recorded on the step result so the handler
  can read them

+++ packages/core/src/Spec/Action/RetryAction.php
@@ -12,7 +12,7 @@ final readonly class RetryAction extends FailureAction
     /** @param list<SuccessCriterion> $criteria */
     public function __construct(
         string $name,
-        public ?int $retryAfter,
+        public ?float $retryAfter,
         public ?int $retryLimit,
         public ?string $stepId,
         public ?string $workflowId,
